Fix assersion

+++ substrate/primitives/runtime-interface/src/impls.rs
@@ -30,8 +30,8 @@ use sp_wasm_interface::{FunctionContext, Result};
 // Make sure that our assumptions for storing a pointer + its size in `u64` is valid.
 #[cfg(all(substrate_runtime, not(feature = "disable_target_static_assertions")))]
 const _: () = {
-	assert!(core::mem::size_of::<usize>() == core::mem::size_of::<usize>());
-	assert!(core::mem::size_of::<*const u8>() == core::mem::size_of::<usize>());
+	assert!(core::mem::size_of::<usize>() == core::mem::size_of::<u32>());
+	assert!(core::mem::size_of::<*const u8>() == core::mem::size_of::<u32>());
 };
 
 /// Implement the traits for the given primitive traits.
